+++ client/search-ui/src/results/sidebar/SearchSidebar.tsx
@@ -19,6 +19,7 @@ import { SectionID } from '@sourcegraph/shared/src/settings/temporary/searchSide
 import { TemporarySettings } from '@sourcegraph/shared/src/settings/temporary/TemporarySettings'
 import { useTemporarySetting } from '@sourcegraph/shared/src/settings/temporary/useTemporarySetting'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
+import { Typography } from '@sourcegraph/wildcard'
 
 import { getDynamicFilterLinks, getRepoFilterLinks, getSearchSnippetLinks } from './FilterLink'
 import { getFiltersOfKind, useLastRepoName } from './helpers'
@@ -180,7 +181,7 @@ export const SearchSidebar: React.FunctionComponent<React.PropsWithChildren<Sear
                         noResultText={
                             <span>
                                 None of the top {repoFilterLinks.length} repositories in your results match this filter.
-                                Try a <code>repo:</code> search in the main search bar instead.
+                                Try a <Typography.Code>repo:</Typography.Code> search in the main search bar instead.
                             </span>
                         }
                     >