+++ client/vscode/src/webview/sidebars/auth/AuthSidebarView.tsx
@@ -1,19 +1,26 @@
 import React, { useMemo, useState } from 'react'
 
+import { VSCodeButton } from '@vscode/webview-ui-toolkit/react'
 import classNames from 'classnames'
 
 import { Form } from '@sourcegraph/branded/src/components/Form'
 import { LoaderInput } from '@sourcegraph/branded/src/components/LoaderInput'
 import { currentAuthStateQuery } from '@sourcegraph/shared/src/auth'
 import { CurrentAuthStateResult, CurrentAuthStateVariables } from '@sourcegraph/shared/src/graphql-operations'
-import { Alert, Typography } from '@sourcegraph/wildcard'
+import { Alert, Typography, Button, Link } from '@sourcegraph/wildcard'
 
+import {
+    VSCE_LINK_DOTCOM,
+    VSCE_LINK_MARKETPLACE,
+    VSCE_LINK_AUTH,
+    VSCE_LINK_TOKEN_CALLBACK,
+    VSCE_LINK_TOKEN_CALLBACK_TEST,
+    VSCE_LINK_USER_DOCS,
+    VSCE_SIDEBAR_PARAMS,
+} from '../../../common/links'
 import { WebviewPageProps } from '../../platform/context'
 
 import styles from './AuthSidebarView.module.scss'
-
-const SIDEBAR_UTM_PARAMS = 'utm_medium=VSCODE&utm_source=sidebar&utm_campaign=vsce-sign-up&utm_content=sign-up'
-
 interface AuthSidebarViewProps
     extends Pick<WebviewPageProps, 'extensionCoreAPI' | 'platformContext' | 'instanceURL' | 'authenticatedUser'> {}
 
@@ -31,9 +38,19 @@ export const AuthSidebarView: React.FunctionComponent<React.PropsWithChildren<Au
     const [state, setState] = useState<'initial' | 'validating' | 'success' | 'failure'>('initial')
     const [hasAccount, setHasAccount] = useState(!authenticatedUser)
     const [usePrivateInstance, setUsePrivateInstance] = useState(false)
-    const signUpURL = `https://sourcegraph.com/sign-up?editor=vscode&${SIDEBAR_UTM_PARAMS}`
+    const signUpURL = VSCE_LINK_AUTH('sign-up')
     const instanceHostname = useMemo(() => new URL(instanceURL).hostname, [instanceURL])
     const [hostname, setHostname] = useState(instanceHostname)
+    const isSourcegraphDotCom = useMemo(() => {
+        const hostname = new URL(instanceURL).hostname
+        if (hostname === 'sourcegraph.com' || hostname === 'www.sourcegraph.com') {
+            return VSCE_LINK_TOKEN_CALLBACK
+        }
+        if (hostname === 'sourcegraph.test') {
+            return VSCE_LINK_TOKEN_CALLBACK_TEST
+        }
+        return null
+    }, [instanceURL])
 
     const validateAccessToken: React.FormEventHandler<HTMLFormElement> = (event): void => {
         event.preventDefault()
@@ -74,10 +91,9 @@ export const AuthSidebarView: React.FunctionComponent<React.PropsWithChildren<Au
         // If successful, update setting. This form will no longer be rendered
     }
 
-    const onSignUpClick = async (): Promise<void> => {
+    const onSignUpClick = (): void => {
         setHasAccount(true)
         platformContext.telemetryService.log('VSCESidebarCreateAccount')
-        await extensionCoreAPI.openLink(signUpURL)
     }
 
     if (state === 'success') {
@@ -108,17 +124,23 @@ export const AuthSidebarView: React.FunctionComponent<React.PropsWithChildren<Au
                     multiple repositories & commit history, monitor code changes, save searches, and more.
                 </p>
                 <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
-                    <button type="button" onClick={onSignUpClick} className={classNames('btn my-1', styles.ctaButton)}>
+                    <Button
+                        variant="primary"
+                        className={classNames(styles.ctaButton, 'btn-text-link p-0 text-left my-1')}
+                        onClick={onSignUpClick}
+                        as={Link}
+                        to={signUpURL}
+                    >
                         Create an account
-                    </button>
+                    </Button>
                 </p>
-                <button
+                <VSCodeButton
                     type="button"
                     className={classNames(styles.ctaParagraph, 'btn btn-text-link text-left')}
                     onClick={() => setHasAccount(true)}
                 >
                     Have an account?
-                </button>
+                </VSCodeButton>
             </>
         )
     }
@@ -131,16 +153,34 @@ export const AuthSidebarView: React.FunctionComponent<React.PropsWithChildren<Au
             <p className={classNames(styles.ctaParagraph)}>
                 See our {/* eslint-disable-next-line react/forbid-elements */}{' '}
                 <a
-                    href={`https://docs.sourcegraph.com/cli/how-tos/creating_an_access_token?${SIDEBAR_UTM_PARAMS}`}
+                    href={VSCE_LINK_USER_DOCS}
                     onClick={() => platformContext.telemetryService.log('VSCESidebarCreateToken')}
                 >
                     user docs
                 </a>{' '}
                 for a video guide on how to create an access token.
             </p>
+            {/* TODO: UPDATE LINK BACK TO CLOUD */}
+            {isSourcegraphDotCom && (
+                <p className={classNames(styles.ctaParagraph)}>
+                    <Link to={isSourcegraphDotCom}>
+                        <VSCodeButton
+                            type="button"
+                            className={classNames(
+                                'btn my-1 p-0',
+                                styles.ctaButton,
+                                styles.ctaButtonWrapperWithContextBelow
+                            )}
+                            autofocus={false}
+                        >
+                            Continue in browser
+                        </VSCodeButton>
+                    </Link>
+                </p>
+            )}
             <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
                 <LoaderInput loading={state === 'validating'}>
-                    <label htmlFor="access-token-input">Access Token</label>
+                    <Typography.Label htmlFor="access-token-input">Access Token</Typography.Label>
                     <input
                         className={classNames('input form-control', styles.ctaInput)}
                         id="access-token-input"
@@ -157,7 +197,7 @@ export const AuthSidebarView: React.FunctionComponent<React.PropsWithChildren<Au
             {usePrivateInstance && (
                 <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
                     <LoaderInput loading={state === 'validating'}>
-                        <label htmlFor="instance-url-input">Sourcegraph Instance URL</label>
+                        <Typography.Label htmlFor="instance-url-input">Sourcegraph Instance URL</Typography.Label>
                         <input
                             className={classNames('input form-control', styles.ctaInput)}
                             id="instance-url-input"
@@ -172,43 +212,45 @@ export const AuthSidebarView: React.FunctionComponent<React.PropsWithChildren<Au
                     </LoaderInput>
                 </p>
             )}
-            <button
+            <VSCodeButton
                 type="submit"
                 disabled={state === 'validating'}
-                className={classNames('btn my-1', styles.ctaButton, styles.ctaButtonWrapperWithContextBelow)}
+                className={classNames('btn my-1 p-0', styles.ctaButton, styles.ctaButtonWrapperWithContextBelow)}
             >
                 Authenticate account
-            </button>
+            </VSCodeButton>
             {state === 'failure' && (
                 <Alert variant="danger" className={classNames(styles.ctaParagraph, 'my-1')}>
                     Unable to verify your access token for {hostname}. Please try again with a new access token.
                 </Alert>
             )}
             {!usePrivateInstance ? (
-                <button
-                    type="button"
-                    className={classNames(styles.ctaParagraph, 'btn btn-text-link text-left my-0')}
+                <Button
+                    variant="link"
+                    className={classNames(styles.ctaParagraph, 'btn-text-link p-0 text-left my-0')}
                     onClick={() => setUsePrivateInstance(true)}
                 >
                     Need to connect to a private instance?
-                </button>
+                </Button>
             ) : (
-                <button
-                    type="button"
-                    className={classNames(styles.ctaParagraph, 'btn btn-text-link text-left my-0')}
+                <Button
+                    variant="link"
+                    className={classNames(styles.ctaParagraph, 'btn-text-link p-0 text-left my-0')}
                     onClick={() => setUsePrivateInstance(false)}
                 >
                     Not a private instance user?
-                </button>
+                </Button>
             )}
             <div className="my-0">
-                <button
-                    type="button"
-                    className={classNames(styles.ctaParagraph, 'btn btn-text-link text-left my-0')}
+                <Button
+                    variant="link"
+                    className={classNames(styles.ctaParagraph, 'btn-text-link p-0 text-left my-0')}
                     onClick={onSignUpClick}
+                    as={Link}
+                    to={signUpURL}
                 >
                     Create an account
-                </button>
+                </Button>
             </div>
         </>
     )
@@ -237,7 +279,7 @@ export const AuthSidebarCta: React.FunctionComponent<React.PropsWithChildren<Aut
                 <p className="mb-0">Learn more:</p>
                 {/* eslint-disable-next-line react/forbid-elements */}
                 <a
-                    href={'https://sourcegraph.com/?' + SIDEBAR_UTM_PARAMS}
+                    href={VSCE_LINK_DOTCOM + VSCE_SIDEBAR_PARAMS}
                     className="my-0"
                     onClick={() => onLinkClick('Sourcegraph')}
                 >
@@ -245,11 +287,7 @@ export const AuthSidebarCta: React.FunctionComponent<React.PropsWithChildren<Aut
                 </a>
                 <br />
                 {/* eslint-disable-next-line react/forbid-elements */}
-                <a
-                    href="https://marketplace.visualstudio.com/items?itemName=sourcegraph.sourcegraph"
-                    className="my-0"
-                    onClick={() => onLinkClick('Extension')}
-                >
+                <a href={VSCE_LINK_MARKETPLACE} className="my-0" onClick={() => onLinkClick('Extension')}>
                     Sourcegraph VS Code extension
                 </a>
             </div>