+++ client/web/src/search/home/SelfHostInstructions.tsx
@@ -78,7 +78,7 @@ export const SelfHostInstructions: React.FunctionComponent<React.PropsWithChildr
                     >
                         <Icon as={ContentCopyIcon} />
                     </Button>
-                    <code className={styles.codeBlock}>{dockerCommand}</code>
+                    <Typography.Code className={styles.codeBlock}>{dockerCommand}</Typography.Code>
                 </MarketingBlock>
                 <div className="d-flex justify-content-between">
                     <Link