+++ internal/database/migration/cliutil/drift.go
@@ -9,7 +9,9 @@ import (
 	"github.com/sourcegraph/sourcegraph/lib/output"
 )
 
-func Drift(commandName string, factory RunnerFactory, outFactory func() *output.Output) *cli.Command {
+type ExpectedSchemaFactory func(repoName, version string) (descriptions.SchemaDescription, error)
+
+func Drift(commandName string, factory RunnerFactory, outFactory OutputFactory, expectedSchemaFactory ExpectedSchemaFactory) *cli.Command {
 	schemaNameFlag := &cli.StringFlag{
 		Name:     "db",
 		Usage:    "The target `schema` to compare.",
@@ -26,22 +28,22 @@ func Drift(commandName string, factory RunnerFactory, outFactory func() *output.
 		if err != nil {
 			return err
 		}
-
 		schemas, err := store.Describe(ctx)
 		if err != nil {
 			return err
 		}
 		schema := schemas["public"]
 
-		expected, err := fetchSchema(schemaNameFlag.Get(cmd), versionFlag.Get(cmd))
+		filename, err := getSchemaJSONFilename(schemaNameFlag.Get(cmd))
+		if err != nil {
+			return err
+		}
+		expectedSchema, err := expectedSchemaFactory(filename, versionFlag.Get(cmd))
 		if err != nil {
 			return err
 		}
 
-		canonicalize(schema)
-		canonicalize(expected)
-
-		return compareSchemaDescriptions(out, schema, expected)
+		return compareSchemaDescriptions(out, canonicalize(schema), canonicalize(expectedSchema))
 	})
 
 	return &cli.Command{
@@ -56,12 +58,14 @@ func Drift(commandName string, factory RunnerFactory, outFactory func() *output.
 	}
 }
 
-func canonicalize(schema descriptions.SchemaDescription) {
-	descriptions.Canonicalize(schema)
+func canonicalize(schemaDescription descriptions.SchemaDescription) descriptions.SchemaDescription {
+	descriptions.Canonicalize(schemaDescription)
 
-	for i, table := range schema.Tables {
+	for i, table := range schemaDescription.Tables {
 		for j := range table.Columns {
-			schema.Tables[i].Columns[j].Index = -1
+			schemaDescription.Tables[i].Columns[j].Index = -1
 		}
 	}
+
+	return schemaDescription
 }